Emit values to Statsd as gauges instead of counters.
This is a workaround on Storm's imprecise metric emit interval, which
oscillates wildly around the user-defined metric emit interval value.

As Storm could (and does) emit multiple metrics in a single defined time
window, Statsd counts would be aggregated by Statsd (and of course,
duplicated), but gauges don't aggregate, so they're a safe alternative.

It would be best to make the Statsd metric type emitted
user-configurable.